Merge 'monitoring' root into 'environment' root
They both do conceptually similar things, and have some overlap. It
makes sense to have them both in one place.

Monitoring configuration is now a child module of the 'environment'
module, and the 'smoke-test' module is now a child of 'smoke-test'.

Another branch and pull request will briefly remove the monitoring
root fully so that this commit can cause it to be recreated.