Use semantically correct "time" tag instead of abbr
We want to show nicely formatted, human-friendly times and dates,
whilst retaining the ability to find the exact timestamp. We were
using te "abbr"eviation tag previously, but it should have been the
"time" tag.

The "title" attribute is included alongside the "datetime" attribute
so that users can mouse over the date and get the proper timestamp
natively.