Dev pipelines trigger on "(dev|prod)-*"
Dev pipelines need to be triggered when a new container image with the format
"prod-*" exists (meaning "this was OK in production"). They should also trigger
whe na new container image with the format "dev-*" exists, so that we can
deploy development versions without confusion.

EventBridge wildcards are NOT regular expressions, so we can't provide a
regular expression like "(dev|prod)-.*" to match bothe prefixes. Instead, we
have to provide two wildcard expressions.

In EventBridge a matcher like {"a": [{"wildcard": "a*"}, {"wildcard": "b*"}]}
will act as a logical OR.

To make it possible to have two prefixes for one pipeline and 1 everywhere
else, we make the config take an array.