Make application pipeline execution mode configurable
We were previously using QUEUED mode by default because we felt this best suited our needs.

However we now plan to have the dev env automatically deploy when a successful
production deploy happens. When we want to avoid this happening we plan to
pause the dev pipeline. Under QUEUED mode this would result in a backlog of
executions building up, which is not what we want to happen. Instead we'd
prefer the behaviour of SUPERSEDED, but only in dev.

Staging, production, and user research environments all continue to use QUEUED mode.