Implement paused pipeline detection
We expect that we will routinely pause pipelines, especially in dev
environments, and it's important that we don't forget to unpause them.

For that purpose we implement a small piece of code that looks for pipelines
with stages whose inbound transition has been paused for longer than some
number of hours. Currently it is set to 24 hours.

If it finds any, it alerts us on Slack.