Deploy paused pipeline detection as a Lambda function
The function is scheduled to run every 12 hours EventBridge rule.

We chose to use AWS Lambda here because a small piece of code run on a schedule
is the type of use case which is AWS Lambda's strong suit. The alternative
would be to use AWS ECS, but that would require
* a container definition
* a pipeline to build and push the container
* pipelines to deploy the container

Which is a lot of extra code for what is otherwise a simple task.