Allow all roles from the environment accounts to publish messages to Slack
We previously only allowed a small number of AWS services to publish messagesn
to the SNS topic which goes out to Slack. We now need at least one additional
role to be able to do the same.

We've chosen to use the root ARN here instead of listing each and every role because
1. we trust those accounts
2. we don't want a circular dependency between the deploy and environment accounts