Remove .ruby-version file for Ruby Lambda functions
We don't actually want to pick a specific version of Ruby when the
pipeline-invoker is run in the context of an AWS Lambda function; the
base image already installs a couple of versions of Ruby [[1]] and lets
you choose the major version by specifying a runtime [[2]].

If we keep a `.ruby-version` file for the Lambda function code we
either have to keep it in sync with the versions included in the Lambda
base images, or add extra time to the function startup installing the
specific version.

Instead this commit just removes the `.ruby-version` file and relies on
the Lambda runtime defined in Terraform [[3], [4]].

When running `make ENV forms/pipelines COMMAND` locally the
`.ruby-version` file in the repository root will be used; as long as we
are keeping the runtime version in Terraform in the same release series
of Ruby this should be fine.

[1]: https://github.com/aws/aws-codebuild-docker-images/blob/9fcc604fde570fa9ede9dfcfb1ab4c7ac093ca36/al2/x86_64/standard/5.0/Dockerfile#L270-L293
[2]: https://github.com/aws/aws-codebuild-docker-images/blob/9fcc604fde570fa9ede9dfcfb1ab4c7ac093ca36/al2/x86_64/standard/5.0/runtimes.yml#L139-L157
[3]: infra/deployments/forms/pipelines/pipeline-execution-lambda.tf
[4]: infra/deployments/forms/pipelines/paused-pipeline-detection.tf

Co-authored-by: Andy Hunt <andy.hunt@digital.cabinet-office.gov.uk>