Ignore local .ruby-version when running on AWS
The `self-update-pipelines` stage in the `apply-forms-terraform-*`
pipelines has the whole forms-deploy repo checked out when running,
which means that without a local `.ruby-version` file in the
`pipeline-invoker` directory (see previous commit) it uses the
`.ruby-version` file in the repo root. This results in our
`.ruby-version` file being tied to what versions of Ruby are available
in the default ECS containers, which we don't want.

Instead this commit works around this by telling rbenv to use the
globally set version of Ruby when running in an AWS container, ignoring
any local `.ruby-version` files.