Improve speed of Terraform validation
To validate our Terraform syntax, we unfortunately need to initialise the root
module before running "terraform validate". Both of these are slow operations,
so we want to do them as little as possible.

Hashicorp maintains a tool called terraform-config-inspect which can parse
Terraform files and tell you about them, including what modules they use. We
can use this to recursively inspect what modules are used in each Terraform
root, and thereby work out which modules and roots need to have their syntax
re-validated.

The tool does not need the root to be initialised, which speeds the whole thing
up a lot.