Use /up endpoint in Route53 health checks
Changes the health check type from HTTP_STR_MATCH to HTTPS. The /up
endpoint does not return any text, only an HTTP 200/500 and a
green/red screen.

According to [1], the HTTPS type will consider an endpoint healthy
if it returns a 2xx or 3xx response within 2 seconds. This matches
the behaviour of the endpoint.

[1] https://docs.aws.amazon.com/Route53/latest/DeveloperGuide/dns-failover-determining-health-of-endpoints.html