Move end-to-end tests to the same pipeline stage as deployment
Because CodePipeline allows stage N-1 to run with a new set of inputs whilst
stage N is running, is it possible in that our end-to-end tests were testing
changes that were being deployed concurrently. This is not what we wanted.

To prevent that happening, we now run the end-to-end tests in the same stage as
the container deployment.