Force environment module to forget alerts, monitoring and shield resources
We're going to move the alerts, monitoring, and shield resources to a new Terraform
root. We don't want Terraform to destroy and re-create them, so that we don't
have a gap in alert coverage. Instead we want Terraform to simply forget them,
so we can import them in a later commit.