Fix email address domain for MOD Directorate of Defence Communications
In commit b63e6b45 we added a list of allowed email address domains,
based on a list in [[1]], which included `ddc-mod.uk`. However, in
unrelated communications later on, we were provided a `ddc-mod.org`
email address for a MOD user [[2]].

There are no `ddc-mod` email addresses in our production user database,
either with `.org` or `.uk`, but it appears this might be a typo.

Searching publicly available information is not totally conclusive, but
does suggest that `ddc-mod.org` email addresses do exist [[3], [4]],
while there is no evidence for `ddc-mod.uk` email addresses [[5]], and
the `ddc-mod.uk` domain does not appear to be registered [[6]]]. Sending
an email to any `ddc-mod.uk` email address results in a "domain not
found error".

I also tried searching our logs for instances of `ddc-mod` email
addresses (from attempted logins), but found nothing, suggesting no
users have been affected by this issue.

As this is most likely a typo, this commit updates our configuration
accordingly.

[1]: https://trello.com/c/5ThqPqfc/
[2]: https://docs.google.com/document/d/1EGHH_jx6bhCQIO3rADDjGHH9ybRsj6GDwrQ7YPLuqdk/edit?disco=AAABOMr48q0
[3]: https://assets.publishing.service.gov.uk/media/5a83025740f0b62305b95368/UseOfMODInsigniaOnPublicMemorialsOrHeadstones-V2-2.pdf
[4]: https://civilservice.blog.gov.uk/2016/12/20/what-the-armed-forces-covenant-means-for-civil-servants/
[5]: https://www.google.com/search?q=%22ddc-mod.uk%22
[6]: https://who.is/whois/ddc-mod.uk