Remove unnecessary "import", "moved", and "removed" blocks.
Once they have been run once and had the desired effect on the state file,
they're no longer necessary and we can remove them.

Some we leave in place as evidence that the resource was not created by
Terraform in the first place, and we don't expect it to be in the future.