Reduce the Auth0 login session timeout
Reduce the absolute timeout before users need to log in again to 8
hours.

We are reducing this to 8 hours due to the following points in our
risk treatment plan:

RTP ref 7:

For Super Administrators accessing the Forms Administration Interface,
 implement the shortest tolerable idle session timeout and absolute
 session timeout.

RTP ref 3:

For Editors accessing the Forms Administration Interface, implement
the shortest tolerable idle session timeout and absolute session
timeout.

We have determined that 8 hours is the shortest tolerable time when
balancing user experience as we expect users to be able to user the
platform for a standard working day without interruption.