Move us-east-1 Zendesk SNS topic to forms/health root
The Zendesk SNS topic in us-east-1 is only ever used by the forms/health root,
so living in forms/environment didn't make much sense.

It has been brought over in such a way that future commits can extract the
eu-west-2 topics out of the alerts module and up to the forms/health root,
where they can all be made consistent and the ARNs given to the modules as
needed.