Move us-east-1 Zendesk alerts topic within Terraform state
After we've refactored the code for the alerts topic, we need to ensure
Terraform will do the correct thing with the state file. To achieve this, we
use a "moved" block to indicate to Terraform that an removed resource and a new
resource point to the same thing in the real world.