Pass listener priority into ecs-service instead of hardcoding
The ecs-service module is used by a number of Terraform roots, which
makes it impossible for us to use Terraform alone to ensure that
listener priorities are distinct.

To assist with that, we add a new custom checkov check:
CKV2_FORMS_AWS_8. This check does its best to ensure that every use
of aws_lb_listener_rule witin the same directory receives only
unique values for listener_priority. It uses the lexical value of
listener_arn to disambiguate between different listeners within the
same directory, but this is not a guarantee.

The check is intended to help us not make an easy mistake, not
completely prevent it from happening.