Add docker registry credentials to parameter store (terraform)
These credentials are only used in the `deploy` account when pulling images. If
we want to, we can create different ones for different environments - this would
be useful if we're hitting the limit of how many images we pull

We also rename the credentials. The word `development` was confusing given that
they're not use for the development environment