add codestar connection as a variable throughout
we were previously hardcoding the codestar connection id throughout the repo.
I've added the codestar_connection_arn as a variable that is passed in through the inputs/tfvar files

This means that if we ever add a new account we can add it to the tfvars, rather than searching where to update it in the code.