refactor github_connection_arn to codestar_connection_arn
previously we have a variable github_connection_arn which was hardcoded to be the deploy account codestar connection arn.
In order to keep naming consistent I have renamed this to codestar_connection_arn.
I have also removed the hardcoded arn from the variable files and am using the passed in arn from the tfvars files instead.