Create an automatic snapshot when we run the forms/rds root
This is a temporary addition. We are planning to upgrade our cluster from Aurora
v1 to Aurora v2, we need to have a snapshot that's as recent as possible. To
make it simpler for ourselves this automatically creates a snapshop whenever the
Terraform runs for the `forms/rds` root

Checkov suggests making sure the snapshot is encrypted. However, when trying to
add "storage_encrypted = true" terraform says it can't configure it and "its
value will be decided automatically based on the result of applying this
configuration". We believe the snapshot will have the same configuration as the
existing cluster which is already encrypted