Use list(number) type for form IDs terraform variable
Declare the "csv_submission_enabled_for_form_ids" forms-runner
variable with type list(number) so that it only accepts values of
the expected type.

forms-runner expects the environment variable to be a comma separated
list of IDs, so use the terraform `join` function to convert the
list of numbers to a comma separated string when setting the
environment variable.

Co-authored-by: Jamie Wilkinson <jamie.wilkinson@digital.cabinet-office.gov.uk>