Allow support role to put up maintenance page
Previously the only ways to put up a maintenance page were by pull request, or by an admin role in an environment.
Since we are now moving to OOH support and there might not always be someone available to approve a PR, and because
most of the tech team do not have access to the admin roles, we have decided to add permissions to put a maintenance page
to the support roles.