Change type for csv_submission_enabled_for_form_ids variable
Change the type from list(number) -> list(string). This is because
we are in the process of giving forms external IDs that are strings
and exposing these publicly rather than the internal database ID.

This means that this feature flag will need to be updated to use
external IDs rather than internal IDs. For existing forms, the
external ID will be equal to the internal ID converted to a string.

This means that we won't need to change the values in this array for
existing forms, we just need to allow for the values to be strings.
The value we pass to the environment variable to forms runner is a
String of comma separate form IDs, so this won't need to change to
support external IDs.