Add a new CSV submissions bucket to deploy/tools
We are experimenting with sending CSV files to an S3 bucket as an
alternative form submission type. In production the S3 bucket will
be owned by the owner of the form, but in development and staging we
need a bucket that we can treat as if its owned by somebody else
(hence putting it in a different account).

The bucket policy is slightly complicated for two reasons:

1. We don't want to encode the names of all the environments, and the names of
   the environments are included in the role name.

2. You can't use a wildcard in the principal section of the IAM policy.

To get around those problems we specify the root principal for the account,
then use a condition to limit it to roles which have a prefix up to where the
environment name is appended.

This commit also contains a small refactor to extract the ids of
other accounts to somewhere reusable.