Add Makefile target for running all Terraform in the correct order
The target reads the running order YAML file to discover the order, then runs them one by one.

There is a lot of output coming from Terraform, so the output is
alsowritten to a timestamped logs directory, with one file per
Terraform root. The files are numerically ordered to match the
sequence of Terraform roots that were run.