Create DynamoDB table in init-state-bucket script; enable state file locking
The `init.sh` script in `infra/deployments/account/init-state-bucket/` is
used to bootstrap an AWS S3 bucket for storing Terraform state files.

We now want to add state file locking to Terraform, which uses a DynamoDB
table. A previous commit implemented the new table in the account root but
this created a chicken and egg situation: the account root wanted to lock
its state file, but the DynamoDB table didn't exist yet, and it couldn't
create it because it couldn't lock its state file.

The best solution is to create the DynamoDB table alongside the state
bucket, so that it's available when the account root and every subsequent
root runs.

IMPORTANT! This commit also, incidentally, enables state file locking for
all Terraform roots. It does this by adding the `dynamodb_table` key to
the Terraform vars files in `infra/deployments/account/tfvars/backends/`.