Correctly specify the Terraform root name in CodeBuild
CodeBuild does not support slashes ("/") in project names, but our roots
contain them. To work around that, we replaced slashes with underscores ("_"),
but that resulted in CodeBuild executng something like `make dev forms_dns
apply`.

To fix THAT, we re-replace underscores with slashes with specifying the root
name for CodeBuild.