Retire Aurora v1 database
We have a new Aurora v2 database which has been in situ and working in
production for 10 days or so without issue. We can now safely delete the old
database.

We can also stop taking snapshots of the database every time we deploy, because
we no longer need those snapshots.