Generate correct deployer role name for the development account
The deploy/ecr root grants the deployer roles in every other
environment access to the ECR repositories. It does this with an IAM
policy which specifies the deployer roles, which it constructs from
the data in the `all-accounts` module.

The account root creates the deployer role, which includes the
`environment_name` variable. In every account bar development, the
environment name is the same as the account name in the
`all-accounts` module. For the dev account the environment name is
"dev" and the account name is "development".

We must account for it specially.