Replace YQ with Sed for setting GHA Terraform version
When we upgrade Terraform version we also set the version constraint
in GitHub Actions. Previously we did a sensible thing and used YQ to
manipulate the YAML file and set the string.

However, there's a bug in the YAML parser that YQ uses which mangles
multi-line strings, which was mangling one of our important scripts.

To work around that we now use Sed to target the specific line in
the file we want to change. This is imperfect.