configure new end to end test
we have added a new end to end test to test the new submission type of s3.
In order to run the test, we need a new role in each of the environments that will submit the form submission to s3 and run the e2e test.
This commit also passes the necessary environment variables to the pipeline.