Increase pipeline-invoker lambda timeout to 10s
We've experienced invocations that take longer than the default 3s timeout.
I don't know why that it, the lambda is stupidly simple. It takes a JSON
blob, transforms it, and makes an AWS CodePipeline API call.

Frankly that shouldn't take 3 actual human seconds, 3000 milliseconds. I'd
be worried if an entire application response with multiple database queries
and whatnot took 1s, nevermind a tiny bit of JSON manipulation and an HTTP
request.