Mark rspec and rubocop as development dependencies
I noticed that RSpec and RuboCop were in the main list of dependencies
for some of our Lambdas, which means they're being installed into the
runtimes. This seems unnecessary, so I've marked them as development and
test dependencies. For completeness I went through all the Gemfiles in
this repo to do the same thing.