Allow deployer role to update managed WAF rules
The deployer role was previously allowed to perform wafv2:UpdateWebACL, but
limited to a small number of WebACLs in the current account.

When we make use of a managed rule group, it appears that AWS creates a
unique copy of the rule group in its own account. Thus, when we made a
change to override part of that rule group, our deployer role was not
permitted to update the ACL, because it didn't exist in the current acount.

This commit adds the discovered ARN pattern to the list of WebACLs that it
is permitted to manage.