Add Puma server to pipeline visualiser
This commit adds the latest version of Puma to the dependencies of the
pipeline visualiser.

`rackup` v2.2.0 removed WEBrick as a dependency [[1]], which we were
relying on to serve the pipeline visualiser web app. It turns out that
`rackup` and WEBrick have been deprecated [[2]], although the Sinatra
framework we use still recommends using `rackup`, with Puma [[3]].

Since WEBrick is no longer recommended for production usage [[1], [4]],
let's use Puma instead, but we keep `rackup` around so we don't need to
change how we run the app.

[1]: https://github.com/rack/rackup/blob/main/releases.md#v220
[2]: https://github.com/rack/rackup?tab=readme-ov-file#soft-deprecation
[3]: https://github.com/sinatra/sinatra/blob/3f6c57780b4f77c1439004a949ea2b3d807423ff/README.md?plain=1#L21
[4]: https://github.com/sinatra/sinatra/commit/4a558503a0ee41f26d4ebc07b478340e8a8a5ed6