ecs-service module: add include_domain_root_listener variable
The ecs-service module can create an ALB listener for `var.root_domain`.
Previously it would do this conditionally, based on whether the application
name was "forms-product-page".

Instead of hard coding that decision, we introduce an input variable.
Callers of the module must now say whether they want that listener to be
created.