Increase paused-pipeline-detector lambda timeout to 10s
We've experienced invocations that take longer than the default 3s
timeout.

We've also seen this in other lambda functions in the past few days
(see 7ecb03d68 for another example). It suggests that something has
changed in AWS Lambda, but as yet we have no news of what.

Let's increase the timeout for now, to avoid getting alarms.