Create new environment-specific SSM parameters for Auth0 creds
The current parameter names for the Auth0 credentials don't contain the
environment name. This makes them apply account-wide. We need them to be
environment specific so each environment can get its own Auth0 tenancy if we
want.

This commit creates the new parameters, but does not set their value. The
value will be set manually in the accounts. Future commits will remove the
old ones and refactor the new ones to not include "_env_specific" in the
resource name.