Use environment specific Auth0 client credentials
In a previous commit we created an environment-specific set of AWS SSM
Parameter Store parameters to hold our Auth0 client credentials. We then let
the commit roll through all environmentsm, and set the value of those
parameters, by hand, to the correct values.

In this commit we begin using the new parameters for Terraform's Auth0
authentication.