Add REXML to forms-cli dependencies
aws-sdk-secretsmanager requires a compatible XML library [[1]], this
commit adds REXML as a direct dependency of forms-cli in the Gemfile.

This was surfacing as an issue before because until RuboCop 1.66 we had
REXML as an indirect dependency [[2]]; however that release removed the
need for an XML parser in RuboCop, so our Gemfile no longer contained a
XML parser compatible with aws-sdk-secretsmanager.

aws-sdk-secretsmanager supports a number of XML parsers, I chose to
stick with REXML because it's portable and we haven't been seeing any
issues with it.

[1]: https://github.com/aws/aws-sdk-ruby/blob/75ae244182e05694989993eef86bacf8fee4b910/gems/aws-sdk-core/lib/aws-sdk-core/xml/parser.rb#L66-L75
[2]: https://github.com/rubocop/rubocop/releases/tag/v1.66.0