ecs-service: redirect root domain to www.root_domain
In the ecs-service module we can a configuration option to add a listener for
the root domain. The listener would serve the same traffic as the subdomain.

In practice this meant we served the same content on www. and on the root
domain. This is bad practice for SEO. Instead you should have an HTTP 301
redirect to the canonical domain. In our case this is www.