Add environment variables for sending submission emails using SES
Add forms-runner environment variables for:
- The configuration set name to use for sending emails. This is the
same on all environments. For now use the existing congfiguration set
that we use for sending Auth0 so we are able to test sending
submission emails using SES, but we will want a separate configuration
set down the line.
- The email address to send submission emails from. This uses the
verified domain "Identity" in SES for each environment.
- The no-reply email address for submission emails. This uses the
verified domain "Identity" in SES for each environment.

We've decided to set both the `from` and `reply-to` email addresses to
be no-reply@domain-for-env as we think this makes it most obvious that
you can't reply to the email.