Remove some unused metrics and reorder
We only need m1-m8 to calculate cpu utilisation for all apps. The rest were
leftover from trying out the dashboard.
We also want them to be in order so that it's a bit more intuitive when looking
at them. This follows the same pattern as the memory utilisation widget