Make "image_tag" variables in Terrraform roots unique
forms-admin, forms-api, forms-product-page, and forms-runner all have a
variable called "image_tag" which is used to override the image tag being used
in the ECS task definition.

This works just fine when running one of the roots, because we can set the
variable using a command line fargument or an environment variable.

However, when we run the script to apply the Terraform roots in order, we're
unable to set the image tag for each application separately, because they all
use the same variable name. To address that problem, we make the names unique.