Change paused pipeline notifications to go to deployment channel for dev
We want to keep the alerts Slack channel for only the most important
notifications; the paused pipeline alerts for the dev environment don't
really fit that criteria, and they've been causing toil recently.

This commit changes the paused pipeline detection module to send the
alerts to the deployment notifications Slack channel if the environment
has the name `dev`.

There's probably a better way to code this configuration, but looking at
how we manage the Slack channel ARNs elsewhere in this repo I think a
more wholesale refactor would be good, so I've left it in this hacky
state for now.