Re-introduce engineer access module to deploy/account
At some point we used the engineer-access module in the deploy/account root. We
know this because we have access to the roles in that account. At a later
point, we removed it, probably when we removed it from a module on which
deploy/account depended. This was not intentional.

This commit re-introduces the engineer-access module to the deploy/account
root, and makes a slight tweak to the module to make it applicable.