Replace engineer-access module usage in integration/account
The engineer-access module has become tailored for forms environments over
time. Instead of spending time and effort somehow refactoring it, we should
integration/account to be driven by the same user config module, but create its
own roles and permissions, appropriate to the integration account.

For now, the support roles have no permissions. We will add those as we
discover what is useful for it to have.