refactor rds module to be more generic
As we are adding another rds cluster, we need this module to be generic. Previously we were assuming only one rds cluster per account and naming things using only the environment name as an identifier. I have introduced a new variable 'identifier' and passed that in so that we can have multiple uses of the module per account.