Change schedule for paused pipeline detector for dev environment
We want to minimise noise in the #govuk-forms-alerts Slack channel; the
paused pipeline alerts for the dev environment have been causing toil
recently because the deploys have been paused for a week.

We've agreed to reduce the frequency of the alerts for the dev and user
research environment, to reduce the amount of noise and toil.