Stop saving bundler config when initing ruby project
We're currently setting values in `.bundle/config` for our Ruby lambda
functions when we could be using environment variables [[1]]. The
advantage of setting environment variables is we don't have to worry
about resetting configuration or affecting the personal configuration
of the local developer. We also don't need the `.bundle/config` file in
the AWS Lambda zip [[2]], so it saves a handful of bytes, and shouldn't
have any effect on the behaviour of the lambdas.

[1]: https://bundler.io/v2.6/man/bundle-config.1.html
[2]: https://docs.aws.amazon.com/lambda/latest/dg/ruby-package.html#ruby-package-dependencies