Make updates for pipeline lambdas trigger only on changes to source
Currently whenever a terraform plan is made it will detect changes are
required to put a new version of the AWS Lambda zip files, even if the
changes have just been applied. This is because we are using the `etag`
and `md5file` to detect when changes are needed, but the ETag for an
object on S3 will not necessarily be an MD5 digest of the data [[1]],
and so the two do not match.

Instead in this commit we change the configuration blocks to use the
`source_hash` argument from the Terraform provider, which stores the
provided digest in the state file. This means that once an apply has
been done for the forms/pipelines deployment no changes will be shown in
a subsequent plan, as expected.

[1]: https://docs.aws.amazon.com/AmazonS3/latest/API/API_Object.html
[2]: https://registry.terraform.io/providers/hashicorp/aws/latest/docs/resources/s3_object#source_hash-1