Allow *-readonly roles to assume the forms-runner ECS task role in dev
Developers need the same set of AWS permissions in local dev as forms-runner
will have when it runs. Rather than maitnain a second set of permissions, we
allow the human *-readonly roles in the dev account to assume the same role as
is assumed by forms-runner in ECS.

This will allow developers to assume-role-chain their way to the same
permissions.