Enable Malware Protection for S3
We want to scan objects uploaded to the "file-upload" bucket for malware. AWS
GuardDuty has a feature (Malware Protection for S3) that allows us to do exactly
that.

At the time of writing, AWS lets you enable this feature independently of
GuardDuty.

This PR includes the permissions needed for scanning, but also to e.g. tag
objects and monitor events related to it.