Do not allow forms-runner to tag objects in the file-upload bucket
We were planning to use tags to determine whether or not a file should be
deleted (e.g. only delete it if it was submitted). But we have decided this is
no longer needed.

We do not want to allow forms-runner to change tags that were created by
GuardDuty Malware Protection becuase forms-runner would be able to mark a file
as "safe" when it isn't

forms-runner should still be allowed to GET tags