Allow all IAM principals in integration account to access ECR repositories
This commit adds a policy to the ECR repositories in the review environemnt
which allows all IAM principals in the AWS account to read and write images.
This is very likely overscoped, however at the time of writing the exact IAM
princiapls involved isn't known. A later commit will reduce the scope.