Rename forms-runner cluster member
Previously I had set the cluster member name to be 'solidqueue', however we are likely to use the new db for more than just the solidQueue data, plus it may be that we switch from solidqueue in the future.
So for clarity I have renamed this cluster member to be {app-name}-{env}-primary.

Note that cluster names must be unique in an environment so we can't use primary as we have for forms-admin and forms-api. This change does not affect the existing forms-api/admin cluster as renaming it would make terraform delete and recreate the cluster.