Remove unneeded move statements
While refactoring the rds module to be more generic, we changed the ssm parameters to be a for_each loop on a list of passed in apps. This caused the existing parameters to add a [0] to their location and therefore moved the location of the parameters. At this point we added moved blocks to let terraform know where the existing parameters had moved to.

Now that this refactor has been deployed and the parameters have been moved, we no longer need these moved blocks, and this commit removes them.