Set default configuration set for from address email identity
Creates a new AWS SES v2 configuration set with identical configuration to the
existing v1 set, used for Auth0.

We name the new set auth0 and set it as the default configuration set to use.
We do this because Auth0 sends OTP emails via SMTP, and it does not set the
headers required to change which configuration set is being used.

Other users of the email identity will need to set the configuration set when
they make the API call.