In production, apply the "latest" tag to successfully deployed container images
We don't use the "latest" tag for deploying images to user-facing environments,
however it is very useful for us to have this for the review apps environment.

The review apps environment uses it to find the current production version of
different container images, so it can deploy a known good copy of components
that aren't being tested or changed. For example, deploying a production
version forms-api when reviewing a change to forms-admin.