Add YAML files in buildspecs paths to list of files to be shellchecked
Previously some CodeBuild buildspecs were not being included in the
Infra as Code checks in our GitHub Actions, because the action assumed
that buildspecs would be named `buildspec.yml`; however they can also be
YAML files with any name in a `buildspecs/**` directory.

This commit tweaks the script to find and lint buildspecs to include
these files.