Enable post-apply scripts in invoke-terraform.sh
The invoke-terraform.sh script already supported pre-init, pre-plan, and
pre-apply scripts, as well as a post-apply script in one instance. This change
expands the support to allow all Terraform roots to provide a post-apply
script. This will be used by integration/review in a subsequent commit.

The source directory is also passed to each script as the third argument.