Create GitHub Actions runner for forms-admin
Creates a CodeBuild-hosted GitHub Actions runner for use by forms-admin. Each
runner must be tied to one repository because it needs to be triggered by an
AWS-managed webhook.

At the time of writing, Terraform doesn't support configuring AWS CodeBuild in
the ways needed for using it as a GitHub Actions runner. To get around that, we
configure it wrongly in Terraform, ignore any future changes that specific bit
of config, and use a post-apply script to ensure it's configured correctly.