Remove read permissions for test submissions to S3 bucket
To write submissions to a third party's S3 bucket, the role assumed by
forms-runner should only need write permissions.

Update the permissions for the bucket in forms-deploy that we use for
end-to-end tests and local testing for the submissions to S3 feature
to only grant write access to match this expectation.