Introduce tflint
tflint is a Terraform linter focussed on, in its own words,

* Finding poissble errors
* Warning about deprecated syntax, unused declarations
* Enforcing best practices, naming conventions

We introduce it for its syntax checking and naming conventions enforcement.

tflint does not replace Checkov. It sits alongside it. Checkov has a strong.
mature  library of cloud provider configuration linting rules which we benefit
a lot from, but it does not have the syntax and naming level linting rules of
tflint.