Make it possible for tflint to be information only
We want to give ourselves time to address any failures identified by tflint, so
we just want them to be informational at first. To support that, we make it
possible to turn that feature on and off in the Makefile. A subsequent will
make use of this to allow the checks to pass regardless in GitHub Actions.