Standardise Terraform resoruce and module naming to snake_case
We have a mix of snake_case and kebab-case naming within our code. We want to
standardise on snake_case.

I wrote a script in F# to

1. parse the output of hcl2json on every Terraform file in the repository
2. find resources and modules whose names aren't snake_case
3. generate Terraform "moved" statements for every resource or module that needs renaming
4. write the Terraform "moved" statements to an appropriately named file

I'm not checking that script into git becuase it's not in our linuga franca,
and it was only needed as a one-off.

After generating the "_moves.tf" files, I went through all the original ".tf"
files and did the renames suggested by the "moved" statements.

This may not have covered every single case of kebab-case naming, but it ought
to be a big step in the right direction.

The PR for this branch will contain the output of Terraform plan for the
affected roots, and any root which uses an affected module.